feat: add native code lens support (#207)
* feat: add native code lens support

This PR add introduces native support for Infracost code lenses. Meaning we can drop our reliance on
the Terraform extension for code symbols. Instead we now use the Infracost metadata to determine the
position and range of the code lens. This should speed up the extension for most users, and improve the
robustness of code lens support.

* fix: bump version

* feat: enable graph evaluator